# web/openapi.json
{
  "components": {
    "schemas": {
      "ActorReference": {
        "additionalProperties": false,
        "description": "Server-authored record, never an external credential.",
        "properties": {
          "actor_id": {
            "minLength": 1,
            "title": "Actor Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "human",
              "system",
              "model"
            ],
            "title": "Kind",
            "type": "string"
          },
          "schema_version": {
            "const": "service-v1",
            "default": "service-v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "kind"
        ],
        "title": "ActorReference",
        "type": "object"
      },
...
      "ApprovalDecision": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "$ref": "#/components/schemas/ActorReference"
          },
          "decided_at": {
            "description": "Server unix seconds; never client-supplied",
            "minimum": 0.0,
            "title": "Decided At",
            "type": "integer"
          },
          "decision": {
            "enum": [
              "approve",
              "return",
              "withdraw"
            ],
            "title": "Decision",
            "type": "string"
          },
          "reason": {
            "default": "",
            "maxLength": 1024,
            "title": "Reason",
            "type": "string"
          },
          "schema_version": {
            "const": "service-v1",
            "default": "service-v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "decision",
          "actor",
          "decided_at"
        ],
        "title": "ApprovalDecision",
        "type": "object"
      },
      "ApprovalDecisionCommand": {
        "additionalProperties": false,
        "description": "approve / return / withdraw, by the authenticated caller, never a named body.",
        "properties": {
          "decision": {
            "enum": [
              "approve",
              "return",
              "withdraw"
            ],
            "title": "Decision",
            "type": "string"
          },
          "idempotency_key": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$",
            "title": "Idempotency Key",
            "type": "string"
          },
          "reason": {
            "default": "",
            "maxLength": 1024,
            "title": "Reason",
            "type": "string"
          },
          "schema_version": {
            "const": "service-v1",
            "default": "service-v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "idempotency_key",
          "decision"
        ],
        "title": "ApprovalDecisionCommand",
        "type": "object"
      },
...
      "ReportApproval": {
        "additionalProperties": false,
        "description": "The durable approval request with its immutable content binding.",
        "properties": {
          "approval_id": {
            "format": "uuid",
            "title": "Approval Id",
            "type": "string"
          },
          "binding": {
            "$ref": "#/components/schemas/ReportVersionBinding"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "payload_digest": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Payload Digest",
            "type": "string"
          },
          "readiness_policy_version": {
            "default": "formal-readiness-v1",
            "title": "Readiness Policy Version",
            "type": "string"
          },
          "run": {
            "$ref": "#/components/schemas/RunReference"
          },
          "schema_version": {
            "const": "service-v1",
            "default": "service-v1",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "enum": [
              "submitted",
              "approved",
              "returned",
              "withdrawn",
              "superseded"
            ],
            "title": "Status",
            "type": "string"
          },
          "submitted_at": {
            "minimum": 0.0,
            "title": "Submitted At",
            "type": "integer"
          },
          "submitted_by": {
            "$ref": "#/components/schemas/ActorReference"
          }
        },
        "required": [
          "approval_id",
          "job_id",
          "run",
          "binding",
          "status",
          "submitted_by",
          "submitted_at",
          "payload_digest"
        ],
        "title": "ReportApproval",
        "type": "object"
      },
...
      "ReportVersionBinding": {
        "additionalProperties": false,
        "description": "The exact content a person approves: one hash per official table.\n\nDeliberately content-only: the snapshot digest, the template bundle and the three\nfilled-workbook hashes. Which readiness policy admitted the submission is recorded\non the approval itself, so two services computing the binding from the same bytes\nalways agree on its identity.",
        "properties": {
          "calculation_snapshot_digest": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Calculation Snapshot Digest",
            "type": "string"
          },
          "schema_version": {
            "const": "service-v1",
            "default": "service-v1",
            "title": "Schema Version",
            "type": "string"
          },
          "template_bundle": {
            "$ref": "#/components/schemas/TemplateBundleReference"
          },
          "workbook_hashes": {
            "additionalProperties": {
              "pattern": "^[a-f0-9]{64}$",
              "type": "string"
            },
            "maxProperties": 3,
            "minProperties": 3,
            "propertyNames": {
              "enum": [
                "table_3",
                "table_4",
                "table_5"
              ]
            },
            "title": "Workbook Hashes",
            "type": "object"
          }
        },
        "required": [
          "calculation_snapshot_digest",
          "template_bundle",
          "workbook_hashes"
        ],
        "title": "ReportVersionBinding",
        "type": "object"
      },
...
      "RevisionReference": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$",
            "title": "Case Id",
            "type": "string"
          },
          "material_digest": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Material Digest",
            "type": "string"
          },
          "revision_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$",
            "title": "Revision Id",
            "type": "string"
          },
          "schema_version": {
            "const": "service-v1",
            "default": "service-v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "revision_id",
          "material_digest"
        ],
        "title": "RevisionReference",
        "type": "object"
      },
...
      "RunReference": {
        "additionalProperties": false,
        "properties": {
          "attempt_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Id"
          },
          "revision": {
            "$ref": "#/components/schemas/RevisionReference"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "runtime_session_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Session Id"
          },
          "schema_version": {
            "const": "service-v1",
            "default": "service-v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "revision"
        ],
        "title": "RunReference",
        "type": "object"
      },
      "ServiceErrorCode": {
        "enum": [
          "invalid_request",
          "unauthorized",
          "not_found",
          "version_conflict",
          "capability_unavailable",
          "execution_failed"
        ],
        "title": "ServiceErrorCode",
        "type": "string"
      },
      "ServiceProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ServiceErrorCode"
          },
          "message": {
            "const": "Service operation could not be completed.",
            "default": "Service operation could not be completed.",
            "title": "Message",
            "type": "string"
          },
          "schema_version": {
            "const": "service-v1",
            "default": "service-v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "ServiceProblem",
        "type": "object"
      },
...
      "SubmitReportApproval": {
        "additionalProperties": false,
        "description": "Untrusted command; the server computes the binding it will be judged by.",
        "properties": {
          "calculation_snapshot_digest": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Calculation Snapshot Digest",
            "type": "string"
          },
          "idempotency_key": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$",
            "title": "Idempotency Key",
            "type": "string"
          },
          "run": {
            "$ref": "#/components/schemas/RunReference"
          },
          "schema_version": {
            "const": "service-v1",
            "default": "service-v1",
            "title": "Schema Version",
            "type": "string"
          },
          "template_bundle": {
            "$ref": "#/components/schemas/TemplateBundleReference"
          }
        },
        "required": [
          "idempotency_key",
          "run",
          "calculation_snapshot_digest",
          "template_bundle"
        ],
        "title": "SubmitReportApproval",
        "type": "object"
      },
...
      "TemplateBundleReference": {
        "additionalProperties": false,
        "description": "A pinned official asset bundle. A request names it; it never carries a path.",
        "properties": {
          "bundle_hash": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Bundle Hash",
            "type": "string"
          },
          "bundle_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$",
            "title": "Bundle Id",
            "type": "string"
          },
          "schema_version": {
            "const": "service-v1",
            "default": "service-v1",
            "title": "Schema Version",
            "type": "string"
          },
          "version": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "bundle_id",
          "version",
          "bundle_hash"
        ],
        "title": "TemplateBundleReference",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Synchronous factor review entry; document extraction requires configuration.",
    "title": "Agentic AI Real Estate Valuation Reviewer",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/review-jobs/{job_id}/report-approvals": {
      "post": {
        "description": "Pin the current content as one report version and submit it for approval.",
        "operationId": "submit_report_approval_v1_review_jobs__job_id__report_approvals_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitReportApproval"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportApproval"
                }
              }
            },
            "description": "Exact replay of a known submission"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportApproval"
                }
              }
            },
            "description": "The version is submitted for approval"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblem"
                }
              }
            },
            "description": "Missing case or publish permission"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblem"
                }
              }
            },
            "description": "No such job or approval"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblem"
                }
              }
            },
            "description": "Not ready, stale content, wrong state, or a reused key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblem"
                }
              }
            },
            "description": "Invalid command"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblem"
                }
              }
            },
            "description": "Approvals are not configured"
          }
        },
        "summary": "Submit Report Approval",
        "tags": [
          "report-approvals"
        ]
      }
    },
    "/v1/review-jobs/{job_id}/report-approvals/{approval_id}": {
      "get": {
        "operationId": "read_report_approval_v1_review_jobs__job_id__report_approvals__approval_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportApproval"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblem"
                }
              }
            },
            "description": "Missing case or publish permission"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblem"
                }
              }
            },
            "description": "No such job or approval"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblem"
                }
              }
            },
            "description": "Not ready, stale content, wrong state, or a reused key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblem"
                }
              }
            },
            "description": "Invalid command"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblem"
                }
              }
            },
            "description": "Approvals are not configured"
          }
        },
        "summary": "Read Report Approval",
        "tags": [
          "report-approvals"
        ]
      }
    },
    "/v1/review-jobs/{job_id}/report-approvals/{approval_id}/decisions": {
      "post": {
        "description": "Approve, return or withdraw - by the authenticated publish-permission holder.",
        "operationId": "decide_report_approval_v1_review_jobs__job_id__report_approvals__approval_id__decisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportApproval"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblem"
                }
              }
            },
            "description": "Missing case or publish permission"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblem"
                }
              }
            },
            "description": "No such job or approval"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblem"
                }
              }
            },
            "description": "Not ready, stale content, wrong state, or a reused key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblem"
                }
              }
            },
            "description": "Invalid command"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblem"
                }
              }
            },
            "description": "Approvals are not configured"
          }
        },
        "summary": "Decide Report Approval",
        "tags": [
          "report-approvals"
        ]
      }
    },
...
  }
}